# apidoc/swagger.yaml
swagger: "2.0"
info:
  version: "1.0"
  title: "Emotion server"
basePath: /emotion/api
description: This is a sample server Petstore server.
consumes:
  - application/json
produces:
  - application/json
paths:

  /user/login:
    get:
      description: User login
      parameters:
      - name: email
        in: path
        type: string
        required: true
      - name: password
        in: path
        type: string
        required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/TokenResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /user/register:
    post:
      description: User register
      parameters:
      - name: email
        in: body
        type: string
        required: true
      - name: password
        in: body
        type: string
        required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/TokenResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /image:
    get:
      description: Get an image (base64 encoded)
      parameters:
      - name: imageName
        in: path
        type: integer
        required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetImageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Add a new image (base64 encoded)
      parameters:
      - name: data
        in: path
        type: string
        required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AddImageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /post:
#    get:
#      description: Get a post
#      parameters:
#      - name: post_id
#        in: body
#        type: integer
#        required: true
#      responses:
#        "200":
#          description: Success
#          schema:
#            $ref: "#/definitions/GetPostResponse"
#        default:
#          description: Error
#          schema:
#            $ref: "#/definitions/ErrorResponse"
    post:
      description: Add new post
      parameters:
      - name: token
        in: body
        type: string
        required: true
      - name: text
        in: body
        type: string
        required: true
      - name: imageName
        in: body
        type: integer
        required: true
      - name: latitude
        in: body
        type: number
        required: true
      - name: longitude
        in: body
        type: number
        required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AddPostResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      description: Get nearby posts
      parameters:
      - name: latitude
        in: path
        type: number
        required: true
      - name: longitude
        in: path
        type: number
        required: true
      - name: range
        in: path
        type: number
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/GetPostResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

# complex objects have schema definitions
definitions:
  TokenResponse:
    required:
      - token
      - token
    properties:
      token:
        type: string
      token:
        type: integer
  ErrorResponse:
    required:
      - error
    properties:
      error:
        type: string
  GetImageResponse:
    required:
      - data
    properties:
      data:
        type: string
  AddImageResponse:
    required:
      - imageName
    properties:
      imageName:
        type: integer
  GetPostResponse:
    required:
      - token
      - text
      - latitude
      - longitude
    properties:
      token:
        type: integer
      text:
        type: string
      imageName:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
  AddPostResponse:
    required:
      - post_id
    properties:
      imageName:
        type: integer
  LocationResponse:
    required:
      - post_id
    properties:
      imageName:
        type: integer